image/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
mangum==0.17.0
boto3==1.26.137
python-dotenv==1.0.0
pydantic==1.10.8
python-jose==3.3.0
python-multipart==0.0.6
pyjwt==2.7.0